requirements.txt
streamlit
playwright
pandas
asyncio
openpyxl
requests>=2.31.0
python-dotenv>=1.0.0
